requirements.txt
streamlit==1.23.1
tensorflow==2.19.0
keras==3.11.1
numpy==1.23.5
pandas==2.0.3
matplotlib==3.7.1
scikit-learn==1.2.2
Pillow==9.4.0
